Rename fixture to helper.
