+++ src/libs/createToolbox/__tests__/createToolbox.test.tsx
@@ -3,13 +3,13 @@ import { Toolbox } from '@/types'
 import '@testing-library/jest-dom'
 import { render, screen } from '@testing-library/react'
 import { createConfiguration } from '../createConfiguration'
-import SayHelloA from './fixtures/SayHelloA.test-fixture'
-import SayHelloB from './fixtures/SayHelloB.test-fixture'
-import SayHelloC from './fixtures/SayHelloC.test-fixture'
-import { ToolboxAComponents } from './fixtures/toolboxA.test-fixture'
-import { ToolboxBComponents } from './fixtures/toolboxB.test-fixture'
-import { ToolboxCComponents } from './fixtures/toolboxC.test-fixture'
-import ToolboxKeys from './fixtures/ToolboxKeys'
+import SayHelloA from './helpers/SayHelloA.test-helper'
+import SayHelloB from './helpers/SayHelloB.test-helper'
+import SayHelloC from './helpers/SayHelloC.test-helper'
+import { ToolboxAComponents } from './helpers/toolboxA.test-helper'
+import { ToolboxBComponents } from './helpers/toolboxB.test-helper'
+import { ToolboxCComponents } from './helpers/toolboxC.test-helper'
+import ToolboxKeys from './helpers/ToolboxKeys.test-helper'
 
 describe('createToolbox', () => {
   describe('Toolbox creation', () => {
