+++ src/libs/createToolbox/__tests__/helpers/SayHelloC.test-helper.tsx
@@ -1,4 +1,4 @@
-import { toolboxCHooks, toolboxCHelpers } from './toolboxC.test-fixture'
+import { toolboxCHooks, toolboxCHelpers } from './toolboxC.test-helper'
 
 export interface ISayHelloProps {}
